Added OpenMP 5.0 specification based lowering of sections
This implementation adds lowering support for sections construct. In the
frontend, a sections construct is further granularized as section blocks
instead of OpenMPConstructs. The constructs, therefore, need special handling in the
Bridge wherein division of executable statements into different section
blocks happens depending on how the statements appear in the source
code.