Fix compile crash with function interface used from modules (#1522)
The interfaceSymbol of an `evaluate::ProcedureDesignator` may be use
or host associated, therefore, it is wrong to directly assume it will
have SubprogramDetails like the previous code was doing in call interface
lowering. `GetUltimate()` must be called first. Add a `getInterfaceDetails`
helper to avoid making fatal assumptions and centralized the code accessing
the `SubprogramDetails` when lowering a caller interface.