Address performance issue from unneeded array copy.
An array copy is made when the array is an argument to an elemental
function. If the function is applied to the entire array value, then the
copy-in/copy-out is elided and any aliasing issues are the user's
responsibility.

Also augments structured FIR ops with newer MLIR interfaces and uses
dependence analysis instead of assuming a particular high-level control
flow structure must be present.

Add regression test.