This fixes a number of issues.
When dealing with values (including aggregates) that have dependent type,
FIR requires that we thread the type parameter values through the IR.
Note that boxed values contain type parameters inside the box. It is a
requirement that the type parameter(s) can be recovered from a box
value.

  - Add verifiers that determine if an Op requires type parameters or
    not and checks that the correct number of parameters is specified.
  - Fix lowering to add the proper number of type parameters.
  - Fix transformations that were losing type parameter information.

Fix code gen bug that was scaling the size of an element twice.

Fix tests to reflect these changes.